Move core logic into subprovider for use in provider-engine
