+++ package.json
@@ -9,7 +9,8 @@
   "author": "kumavis",
   "license": "ISC",
   "dependencies": {
-    "readable-stream": "^2.0.5"
+    "readable-stream": "^2.0.5",
+    "web3-provider-engine": "^12.0.6"
   },
   "devDependencies": {},
   "repository": {
